+++ run/execotest.py
@@ -1,0 +1,5 @@
+#mpyexo
+print('execo test start')
+import c
+c.execo('hello.o')
+print('execo test end')