+++ RELEASE_NOTES.md
@@ -142,4 +142,5 @@
 - mpymod loader now stores scripts in the `env` module so boot doesn't fail when
   builtins are read-only
 - Removed MicroPython mpymod loader again and added support for init/kernel/init.elf
+- New built-in `c.execo()` function runs preloaded .o modules from MicroPython